Dont process get_peers results if there are enough peers for an infohash.
This was unitentionally adding all hosts received via get_peers to the routing
table, which consumes memory. The correct behavior is to respect the
"numTargetPeers" setting provided by the DHT controlling client.

Now whenever we receive the response for get_peers for an infohash that we have
enough peers, just discard the result right away.