Reduce the baseline memory usage for the infohash LRU.
since this is an LRU, items are probably replaced often, therefore it doesn't
make sense to pay the price to allocate all the memory necessary to hold all
the contacts for this infohash. It's unlikely that they will be filled.

That said, this doesn't get rid of the worst case possibility of many
infohashes being half-full of contacts.