dont spam hosts that never replied.
This kind of breaks the engine since it was (accidentally) relying on
these spammy retries to ensure it had a good routing table.

With the present commit, when running in a busy network/server with lots
of packet loss, not even the basic tests will pass because it doesn't
get enough replies from nodes.

It's possible that there is a bug causing this high ratio of sends not
seeing replies, if congested network alone doesn't explain it.

Also renamed lastTime to lastResponseTime.