Avoid panic when calling BinaryToDottedPort over results.
BinaryToDottedPort should not be called when an address has less
than 6 bytes, so do not store such bogus addresses in the
peerContactsSet, and don't return them to the client.

Ideally we should also change BinaryToDottedPort to have an error
argument, but Taipei Torrent depends on it so it should be changed
first.

Fixes issues #16.